feat: move permissionModes to common place
