+++ src/core/utils/index.ts
@@ -38,4 +38,4 @@ export {
 } from './context';
 export { DEFAULT_APP_CATEGORIES } from './default_app_categories';
 export { DEFAULT_WORKSPACE_TEMPLATES } from './default_workspace_templates';
-export { WORKSPACE_PATH_PREFIX } from './constants';
+export { WORKSPACE_PATH_PREFIX, PermissionMode } from './constants';
